+++ app/src/main/java/com/mskd/flux/data/repository/snackbars/SnackbarRepositoryImpl.kt
@@ -1,16 +1,11 @@
 package com.mskd.flux.data.repository.snackbars
 
-import android.content.Context
 import androidx.datastore.core.DataStore
-import androidx.datastore.core.handlers.ReplaceFileCorruptionHandler
 import androidx.datastore.preferences.core.Preferences
 import androidx.datastore.preferences.core.edit
-import androidx.datastore.preferences.core.emptyPreferences
 import androidx.datastore.preferences.core.intPreferencesKey
-import androidx.datastore.preferences.preferencesDataStore
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
-import javax.inject.Inject
 
 class SnackbarRepositoryImpl(val snackbarDataStore: DataStore<Preferences>) : SnackbarRepository {